Move insertion points code under subpackage
