+++ java/src/main/kotlin/io/spine/validation/java/point/BuilderInsertionPoint.kt
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.validation.java
+package io.spine.validation.java.point
 
 import com.google.errorprone.annotations.Immutable
 import com.google.protobuf.GeneratedMessageV3
