+++ java/src/main/kotlin/io/spine/validation/java/point/ExtraValidation.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.point
 
 import io.spine.protodata.ast.TypeName
 import io.spine.protodata.render.CoordinatesFactory.Companion.nowhere
