+++ java/src/main/kotlin/io/spine/validation/java/point/PrintValidationInsertionPoints.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.point
 
 import io.spine.protodata.render.InsertionPointPrinter
 import io.spine.tools.code.Java
